fix issue with existance of "cfg/rcon_passwd.cfg"
Since I use a password which has to be stored in that file and I don't want the whole internet to know my rcon-password, I just told git to ignore the changes I made to that file.
This worked, untill I started switching branches, which overwrote the changes. This in turn broke every rcon feature, since the password was wrong.
Solution: Don't ship the file at all, let the user creat it, remove it from the repo so it can't be overwritten. Hope this works.