fix issue with rcon
Due to the way some text-editors work, a "new line" gets appended at the end of any file. This caused the contents to be misinterpreted to also contain a new-line, which is an incorrect password.
Another change was made to the way rcon was implemented in general. Now a constant connection is maintained.