fix weird half-crash
for some reason, replacing multiprocessing with a regular for-loop fixes a weird issue that was present in .exe builds, where the after some time a lot of gui-instances where started, which didn't respond.
Based on the time-delay of starting the exe and the windows appearing, I guessed it had to do something with translating the messages. Reverting to a regular for-loop did fix the issue, allthough this slows shit down.