refactor code, fix tf2dir error
There where some issues with the way, the code handled not being able to open "console..log". This issue was found alongside some more and I was given feedback for the app's improvement.
The changes include a fix for the tf2dir-issue, variables for colours instead of hard coded colours in the gui and a seperation of gui and backend into seperate files.